techpack: camera: LC898124: Widen DataDM to uint64_t.
../techpack/camera/drivers/cam_sensor_module/cam_ois/onsemi_fw/LC898124/DownloadCmd.c:375:4: warning: cast to smaller integer type 'int' from 'const uint8_t *' (aka 'const unsigned char *') [-Wpointer-to-int-cast]
        , (int)ptr->DataPM , (int)ptr->LengthPM , (int)ptr->Parity , (int)ptr->DataDM , (int)ptr->LengthDMA , (int)ptr->LengthDMB );
          ^~~~~~~~~~~~~~~~
../techpack/camera/drivers/cam_sensor_module/cam_ois/onsemi_fw/LC898124/Ois.h:580:42: note: expanded from macro 'TRACE'
 #define TRACE(...)             CAM_ERR(CAM_OIS, ## __VA_ARGS__)
                                                    ^~~~~~~~~~~
../techpack/camera/drivers/cam_utils/cam_debug_util.h:86:57: note: expanded from macro 'CAM_ERR'
                cam_get_module_name(__module), __func__,  __LINE__, ##args)
                                                                      ^~~~
../include/linux/printk.h:304:33: note: expanded from macro 'pr_err'
        printk(KERN_ERR pr_fmt(fmt), ##__VA_ARGS__)
                                       ^~~~~~~~~~~
../techpack/camera/drivers/cam_sensor_module/cam_ois/onsemi_fw/LC898124/DownloadCmd.c:375:63: warning: cast to smaller integer type 'int' from 'const uint8_t *' (aka 'const unsigned char *') [-Wpointer-to-int-cast]
        , (int)ptr->DataPM , (int)ptr->LengthPM , (int)ptr->Parity , (int)ptr->DataDM , (int)ptr->LengthDMA , (int)ptr->LengthDMB );
                                                                     ^~~~~~~~~~~~~~~~
../techpack/camera/drivers/cam_sensor_module/cam_ois/onsemi_fw/LC898124/Ois.h:580:42: note: expanded from macro 'TRACE'
 #define TRACE(...)             CAM_ERR(CAM_OIS, ## __VA_ARGS__)
                                                    ^~~~~~~~~~~
../techpack/camera/drivers/cam_utils/cam_debug_util.h:86:57: note: expanded from macro 'CAM_ERR'
                cam_get_module_name(__module), __func__,  __LINE__, ##args)
                                                                      ^~~~
../include/linux/printk.h:304:33: note: expanded from macro 'pr_err'
        printk(KERN_ERR pr_fmt(fmt), ##__VA_ARGS__)
                                       ^~~~~~~~~~~

+++ techpack/camera/drivers/cam_sensor_module/cam_ois/onsemi_fw/LC898124/DownloadCmd.c
@@ -374,7 +374,7 @@ unsigned char SelectDownload(UINT8 GyroSelect, UINT8 ActSelect, UINT8 MasterSlav
 
 	// 高速化対応Download
 TRACE("DataPM( %08x ), LengthPM( %08x ) , Parity( %08x ), DataDM( %08x ) , LengthDMA( %08x ) , LengthDMB( %08x ) \n"
-	, (int)ptr->DataPM , (int)ptr->LengthPM , (int)ptr->Parity , (int)ptr->DataDM , (int)ptr->LengthDMA , (int)ptr->LengthDMB );
+	, (uint64_t)ptr->DataPM , (int)ptr->LengthPM , (int)ptr->Parity , (uint64_t)ptr->DataDM , (int)ptr->LengthDMA , (int)ptr->LengthDMB );
 	return( DownloadToEP3( ptr->DataPM, ptr->LengthPM, ptr->Parity, ptr->DataDM, ptr->LengthDMA , ptr->LengthDMB ) ); 
 }
